+++ .env.example
@@ -4,9 +4,5 @@ APP_SECRET=
 APP_WEB_URL=http://localhost:3000
 APP_API_URL=http://localhost:3333
 
-# HERE Credentials
-HERE_BASE_API_URL=
-HERE_URL_REQUEST=
-HERE_ACCESS_KEY=
-HERE_ACCESS_SECRET=
+
