+++ src/modules/addresses/dtos/ICreateAddressDTO.ts
@@ -1,6 +1,6 @@
 export default interface ICreateAddressDTO {
   user_id: string;
-  zip_code: string;
+  zip_code?: string;
   neighborhood?: string;
   city?: string;
   state?: string;
