+++ src/shared/container/index.ts
@@ -1,7 +1,6 @@
 import { container } from 'tsyringe';
 
 import '@modules/users/providers';
-import './providers';
 
 import IUsersRepository from '@modules/users/repositories/IUsersRepository';
 import UsersRepository from '@modules/users/infra/typeorm/repositories/UsersRepository';
