--- src/shared/container/providers/MapProvider/models/IMapProvider.ts
@@ -1,5 +1,0 @@
-import { IAddress } from '@modules/users/services/ImportUsersService';
-
-export default interface IMapProvider {
-  formatAddress(address: IAddress): Promise<IAddress | null>;
-}
