--- src/shared/container/providers/index.ts
@@ -1,6 +1,0 @@
-import { container } from 'tsyringe';
-
-import IMapProvider from './MapProvider/models/IMapProvider';
-import HEREMapProvider from './MapProvider/implementations/HEREMapProvider';
-
-container.registerSingleton<IMapProvider>('MapProvider', HEREMapProvider);
